+++ h3-datagram/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "h3-datagram"
+version = "0.0.1"
+edition = "2021"
+
+[dependencies]
+h3 = { path = "../h3" }
+bytes = "1.4"
+pin-project-lite = { version = "0.2", default_features = false }
