+++ h3-quinn/Cargo.toml
@@ -22,3 +22,4 @@ quinn-proto = { version = "0.10", default-features = false }
 tokio-util = { version = "0.7.7" }
 futures = { version = "0.3.27" }
 tokio = { version = "1.28", features = ["io-util"], default-features = false }
+h3-datagram = { path = "../h3-datagram" }
