+++ h3-webtransport/Cargo.toml
@@ -12,6 +12,7 @@ http = "0.2.9"
 pin-project-lite = { version = "0.2", default_features = false }
 tracing = "0.1.37"
 tokio = { version = "1.28", default_features = false }
+h3-datagram = { path = "../h3-datagram" }
 
 [dependencies.h3]
 version = "0.0.2"
