+++ h3/src/error.rs
@@ -200,7 +200,9 @@ impl Code {
         })
     }
 
-    pub(crate) fn with_cause<E: Into<Cause>>(self, cause: E) -> Error {
+    // Todo: public? private?
+    #[doc(hidden)]
+    pub fn with_cause<E: Into<Cause>>(self, cause: E) -> Error {
         Error::from(self).with_cause(cause)
     }
